relocate classify/ contents: fasttext -> datakit/, all_sources_topic -> cluster/weborganizer/
Promote fasttext helpers up one level so they're discoverable as the
generic datakit fasttext step rather than buried under classify/.
Move the WebOrganizer topic-classifier pipeline into the cluster/
subtree alongside the other weborganizer-shaped work.
